EvolutionParameters fix, Added "Generations" too
Added "Generations" to evolution parameters. Fixed the problem of them
not saving. I also now made it that it loads the evolution parameters
with the default values and then with the updated ones everytime the
user opens them. Now the "Start" menu item runs the "oneGeneration()"
mehod in a for loop that is "Generations" long.  That part is still
buggy.